shop_info_service/requirements.txt
flask
Flask-Smorest
Flask-SQLAlchemy
marshmallow
python-dotenv
gunicorn
Flask-Migrate
psycopg2